Implement loopification for local bindings (#14068)
This is a relatively prelimary version. I am sure there is a huge number
of invariants that this breaks, and conditions that I am not checking
etc. I do not even know if the simplifier is the right place to
implement this.

But it works in this simple case:

    module T14068 where
    foo p f k =
      let bar a = if p a then bar (f a) else a
      in k bar

so we can iterate from here.