Prevent inlining of loopified programs
Previously, a recursive function is not inlineable. After loopification,
it turns into a non-recursive function, and suddenly it is. While this
is in general desirable, it has many knock-on effects, which makes it
hard to evaluate and debug loopification. Therefore, this commit (tries to)
prevent this inlining. When this results in no unfixable regressions,
then we can tackle the next step.